♻️ Refactor connect wallet logic
